app/ml_model.py
# app/ml_model.py
"""
Machine Learning Model Placeholder
Replace these functions with your trained model
"""

class SignLanguageModel:
    def __init__(self):
        """
        TODO: Load your trained model here when ready
        Example:
            import tensorflow as tf
            self.model = tf.keras.models.load_model('models/sign_language_model.h5')
        """
        self.model = None
        print("⚠️  Using DUMMY translation model (replace with real model)")
    
    def translate_video(self, video_path):
        """
        Translates sign language video to text
        
        Args:
            video_path (str): Path to uploaded .webm video
            
        Returns:
            str: Translated English text
            
        TODO: Implement actual translation
        Steps:
            1. Extract frames from video
            2. Preprocess frames (resize, normalize)
            3. Run model inference
            4. Post-process output to text
        """
        # DUMMY IMPLEMENTATION - Replace with real model
        dummy_responses = [
            "I have severe headache and fever since yesterday",
            "My stomach hurts and I feel nauseous",
            "I have chest pain and difficulty breathing",
            "I injured my leg and cannot walk properly",
            "I have persistent cough and sore throat"
        ]
        
        import random
        return random.choice(dummy_responses)
    
    def generate_sign_video(self, text, output_path):
        """
        Generates sign language video from text (FUTURE FEATURE)
        
        Args:
            text (str): Doctor's text response
            output_path (str): Where to save generated video
            
        Returns:
            bool: Success status
            
        NOTE: This is extremely complex and may not be feasible.
              Consider using pre-recorded sign videos instead.
        """
        print(f"⚠️  Video generation not implemented")
        print(f"   Text: {text}")
        print(f"   Output path: {output_path}")
        return False

# Create singleton instance
model = SignLanguageModel()